+++ lib/PointyCastleN/paddings/pkcs7.dart
@@ -14,7 +14,7 @@ import '../src/ufixnum.dart';
 /// A [Padding] that adds PKCS7/PKCS5 padding to a block.
 class PKCS7Padding extends BasePadding {
   static final FactoryConfig FACTORY_CONFIG =
-      new StaticFactoryConfig(Padding, "PKCS7", () => PKCS7Padding());
+      StaticFactoryConfig(Padding, "PKCS7", () => PKCS7Padding());
 
   String get algorithmName => "PKCS7";