+++ lib/src/algos/chacha/chacha12.dart
@@ -6,7 +6,6 @@
 import 'dart:typed_data';
 import 'dart:convert';
 
-import 'package:meta/meta.dart';
 
 /// Implements Chacha20 ([https://tools.ietf.org/html/rfc7539](RFC 7539)), which
 /// has become a popular symmetric stream cipher. It's now used by protocols
@@ -36,8 +35,8 @@ class Chacha12 extends Converter<List<int>, Uint8List> {
 
   /// Sets initial state using the key, nonce, and counter.
   void initialize({
-    @required List<int> key,
-    @required List<int> nonce,
+    List<int> key,
+    List<int> nonce,
     int keyStreamIndex = 0,
   }) {
     if (key == null) {
@@ -285,7 +284,6 @@ class Chacha12 extends Converter<List<int>, Uint8List> {
   ///
   /// Throws [StateError] if [initialize] has not been invoked or [clear] has
   /// been invoked.
-  @visibleForTesting
   void encryptState(Uint32List state) {
     // Validate that we have a proper initial state.
     _validateInitialState(initialState);