+++ lib/src/algos/chacha/chacha8.dart
@@ -1,8 +1,6 @@
 import 'dart:typed_data';
 import 'dart:convert';
 
-import 'package:meta/meta.dart';
-
 /// Implements Chacha20 ([https://tools.ietf.org/html/rfc7539](RFC 7539)), which
 /// has become a popular symmetric stream cipher. It's now used by protocols
 /// such as TLS.
@@ -31,8 +29,8 @@ class Chacha8 extends Converter<List<int>, Uint8List> {
 
   /// Sets initial state using the key, nonce, and counter.
   void initialize({
-    @required List<int> key,
-    @required List<int> nonce,
+    List<int> key,
+    List<int> nonce,
     int keyStreamIndex = 0,
   }) {
     if (key == null) {
@@ -280,7 +278,7 @@ class Chacha8 extends Converter<List<int>, Uint8List> {
   ///
   /// Throws [StateError] if [initialize] has not been invoked or [clear] has
   /// been invoked.
-  @visibleForTesting
+
   void encryptState(Uint32List state) {
     // Validate that we have a proper initial state.
     _validateInitialState(initialState);