+++ lib/src/steel_crypt_base.dart
@@ -27,3 +27,4 @@ part 'pass.dart'; //for PassCrypt
 part 'aes.dart'; //for AesCrypt
 part 'key.dart'; //for CryptKey
 part 'hash.dart'; //for HashCrypt
+part 'mac.dart'; //for both MAC classes