+++ test/steel_crypt_test.dart
@@ -1,5 +1,0 @@
-import 'package:steel_crypt/steel_crypt.dart';
-
-void main() {
-
-}